bugfix(produto): corrigindo problemas de conversão de tipos do double pro bigdecimal nos DTO's de produto

+++ qfome-backend/src/main/java/com/qfome/dto/ProdutoDTO.java
@@ -1,11 +1,13 @@
 package com.qfome.dto;
 
+import java.math.BigDecimal;
+
 import com.qfome.model.Produto;
 
 public record ProdutoDTO(
         Long id,
         String nome,
-        Double preco,
+        BigDecimal preco,
         String slug,
         CategoriaDTO categoria
 ) {
